requirements.txt
elastic-transport==8.13.1
elasticsearch==8.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1